Moved unused files into misc
