+++ DensenetEncoder.py
